// src/enums/dashboard.enum.ts
export enum WIDGET_SIZE {
	SM = 'sm',
	MD = 'md',
	LG = 'lg',
	WIDE = 'wide',
	XL = 'xl',
}

export enum WIDGET_TYPE {
	STAT_KEYS = 'stat-keys',
	STAT_COVERAGE = 'stat-coverage',
	STAT_LANGUAGES = 'stat-languages',
	STAT_UNUSED = 'stat-unused',
	PROJECTS = 'projects',
	LANGUAGES_COVERAGE = 'languages-coverage',
	LAST_ACTIVITY = 'last-activity',
	REVIEW_QUEUE = 'review-queue',
}

export enum DATA_SOURCE_TYPE {
	GLOBAL = 'global',
	PROJECT = 'project',
}